requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
blis==1.2.0
    # via thinc
cachetools==5.5.2
    # via google-auth
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via typer
cloudpathlib==0.21.0
    # via weasel
confection==0.1.5
    # via
    #   thinc
    #   weasel
cymem==2.0.11
    # via
    #   preshed
    #   spacy
    #   thinc
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.24.2
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.164.0
    # via google-generativeai
google-auth==2.38.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.4
    # via -r requirements.in
googleapis-common-protos==1.69.2
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via requests
jinja2==3.1.6
    # via spacy
langcodes==3.5.0
    # via spacy
language-data==1.3.0
    # via langcodes
marisa-trie==1.2.1
    # via language-data
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
murmurhash==1.0.12
    # via
    #   preshed
    #   spacy
    #   thinc
numpy==2.0.2
    # via
    #   blis
    #   spacy
    #   thinc
packaging==24.2
    # via
    #   spacy
    #   thinc
    #   weasel
preshed==3.0.9
    # via
    #   spacy
    #   thinc
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.6
    # via
    #   confection
    #   google-generativeai
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
pyparsing==3.2.1
    # via httplib2
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   google-api-core
    #   spacy
    #   weasel
rich==13.9.4
    # via typer
rsa==4.9
    # via google-auth
shellingham==1.5.4
    # via typer
smart-open==7.1.0
    # via weasel
spacy==3.8.3
    # via -r requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.5.1
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
thinc==8.3.4
    # via spacy
tqdm==4.67.1
    # via
    #   google-generativeai
    #   spacy
typer==0.15.2
    # via
    #   spacy
    #   weasel
typing-extensions==4.12.2
    # via
    #   cloudpathlib
    #   google-generativeai
    #   pydantic
    #   pydantic-core
    #   rich
    #   typer
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.3.0
    # via requests
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.1
    # via spacy
wrapt==1.17.2
    # via smart-open

# The following packages are considered to be unsafe in a requirements file:
# setuptools
